// frontend/src/app/shared/api-models/model/clientLocationModel.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Specifies a client location
 */
export interface ClientLocationModel { 
    /**
     * Client country
     */
    country: string;
    /**
     * Client ip address
     */
    ipAddress: string;
    /**
     * Client location type
     */
    locationType: ClientLocationModel.LocationTypeEnum;
}
export namespace ClientLocationModel {
    export type LocationTypeEnum = 'ANDROID' | 'IOS' | 'WEB';
    export const LocationTypeEnum = {
        ANDROID: 'ANDROID' as LocationTypeEnum,
        IOS: 'IOS' as LocationTypeEnum,
        WEB: 'WEB' as LocationTypeEnum
    };
}